extract "meta" module
in order to reduce amount of exposed private utils via "exports"
