+++ tests/unit/-private/action-test.js
@@ -1,7 +1,7 @@
 import { module, test } from 'qunit';
 import { create } from 'ember-cli-page-object';
 import action from 'ember-cli-page-object/-private/action';
-import { isPageObject } from 'ember-cli-page-object/-private/helpers';
+import { isPageObject } from 'ember-cli-page-object/-private/meta';
 import { setAdapter, getAdapter } from 'ember-cli-page-object/adapters';
 import Adapter from 'ember-cli-page-object/adapter';
